Apply to shared instance
